pending review page
created pending review page
allowed all posts with pending status to be shown.
created a simple template to show the title ( this will need to be updated to show all content)
added in some animation/transitions using css and js to shrink and grow the cell... this will late be used to expand a cell to show content. (allow only one cell to large at a time.